move shared assets one level up
