+++ src/modules/featureInfo/components/featureInfoPanel/FeatureInfoPanel.js
@@ -7,7 +7,7 @@ import { $injector } from '../../../../injection';
 import { abortOrReset } from '../../../../store/featureInfo/featureInfo.action';
 import { AbstractMvuContentPanel } from '../../../menu/components/mainMenu/content/AbstractMvuContentPanel';
 import css from './featureInfoPanel.css';
-import arrowLeftShortIcon from './assets/arrowLeftShort.svg';
+import arrowLeftShortIcon from '../assets/arrowLeftShort.svg';
 import shareIcon from './assets/share.svg';
 import printerIcon from './assets/printer.svg';
 import {
